requirements.txt
dassl.egg==info
ftfy==6.1.3
numpy==1.24.4
numpy==1.24.3
numpy==1.22.4
open_clip_torch==2.24.0
openai==0.27.3
pandas==2.0.3
Pillow==9.4.0
Pillow==10.2.0
regex==2023.3.23
scipy==1.12.0
setuptools==60.2.0
timm==0.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
waffle_utils==1.1.0
yacs==0.1.8
torch==1.13.0+cu116
